refactor: GraphBuilder와 Graph 통합
